requirements.txt
fastapi==0.73.0
motor==2.5.1
pydantic==1.8.2
requests==2.26.0
schedule==1.1.0
uvicorn==0.17.4
pymongo[srv]
python-decouple